4.21*: Blog List Expansion, step 9

+++ controllers/blogs.js
@@ -16,7 +16,7 @@ blogsRouter.get('/', async (request, response) => {
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
-  
+
     const user = await User.findById(decodedToken.id)
 
     if (!body.title || !body.url) {
@@ -40,10 +40,20 @@ blogsRouter.get('/', async (request, response) => {
       response.status(201).json(savedBlog)
   })
 
-  blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
-  })
+ blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'only the creator can delete the blog' })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
+})
 
   blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
